use float precision

+++ src/main/java/top/kzre/pixelshape/Noises.java
@@ -3,15 +3,12 @@ package top.kzre.pixelshape;
 import java.util.Random;
 
 /**
- * 噪声生成工具，提供 Perlin 二维噪声及排列表创建。
+ * 噪声生成工具（float 精度）。
  */
 public final class Noises {
 
     private Noises() {}
 
-    /**
-     * 根据种子创建一个长度为 512 的排列表，供 perlin2D 使用。
-     */
     public static int[] createPermutation(long seed) {
         int[] perm = new int[256];
         for (int i = 0; i < 256; i++) {
@@ -25,30 +22,21 @@ public final class Noises {
         return p;
     }
 
-    /**
-     * Perlin 二维噪声。
-     * @param x x 坐标（通常已缩放）
-     * @param y y 坐标
-     * @param p 排列表，由 createPermutation 生成
-     * @return 噪声值，范围约 [-1, 1]
-     */
-    public static double perlin2D(double x, double y, int[] p) {
+    public static float perlin2D(float x, float y, int[] p) {
         int xi = (int) Math.floor(x) & 255;
         int yi = (int) Math.floor(y) & 255;
-        double xf = x - Math.floor(x);
-        double yf = y - Math.floor(y);
-        double u = fade(xf);
-        double v = fade(yf);
+        float xf = x - (float) Math.floor(x);
+        float yf = y - (float) Math.floor(y);
+        float u = fade(xf);
+        float v = fade(yf);
         int aa = p[p[xi] + yi];
         int ab = p[p[xi] + yi + 1];
         int ba = p[p[xi + 1] + yi];
         int bb = p[p[xi + 1] + yi + 1];
-        double x1 = Utils.lerp(grad(aa, xf, yf), grad(ba, xf - 1, yf), u);
-        double x2 = Utils.lerp(grad(ab, xf, yf - 1), grad(bb, xf - 1, yf - 1), u);
+        float x1 = Utils.lerp(grad(aa, xf, yf), grad(ba, xf - 1, yf), u);
+        float x2 = Utils.lerp(grad(ab, xf, yf - 1), grad(bb, xf - 1, yf - 1), u);
         return Utils.lerp(x1, x2, v);
     }
-
-    // ---- 内部辅助 ----
 
     private static void shuffle(int[] arr, Random rand) {
         for (int i = arr.length - 1; i > 0; i--) {
@@ -59,14 +47,14 @@ public final class Noises {
         }
     }
 
-    private static double fade(double t) {
+    private static float fade(float t) {
         return t * t * t * (t * (t * 6 - 15) + 10);
     }
 
-    private static double grad(int hash, double x, double y) {
+    private static float grad(int hash, float x, float y) {
         int h = hash & 3;
-        double u = h < 2 ? x : y;
-        double v = h < 2 ? y : x;
+        float u = h < 2 ? x : y;
+        float v = h < 2 ? y : x;
         return ((h & 1) == 0 ? u : -u) + ((h & 2) == 0 ? v : -v);
     }
 }
